refactoring of the priority package improve tests
